Adding search definition in header

+++ include/bit_set.h
@@ -6,5 +6,5 @@ public:
     uint8_t BitsSetTable256[256];
     BitSet();
 
-    int countSetBits64(int x);
+    int countSetBits64(uint64_t x);
 };
